# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/cooks": {
      "get": {
        "description": "Cursor pagination (`limit`, `cursor`).",
        "operationId": "CooksController_findAll",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1–100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from the previous page `pagination.nextCursor`",
            "schema": {
              "example": "eyJjcmVhdGVkQXQiOiIyMDI2LTA4LTE2VDEyOjAwOjAwLjAwMFoiLCJpZCI6IjY4YWYifQ",
              "type": "string"
            }
          },
          {
            "name": "lat",
            "required": false,
            "in": "query",
            "schema": {
              "example": 3.8833,
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": false,
            "in": "query",
            "schema": {
              "example": -77.0319,
              "type": "number"
            }
          },
          {
            "name": "radius",
            "required": false,
            "in": "query",
            "description": "Search radius in meters",
            "schema": {
              "default": 10000,
              "example": 10000,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List active cooks (optional proximity: lat, lng, radius meters)",
        "tags": [
          "cooks"
        ]
      },
      "post": {
        "operationId": "CooksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCookDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create cook profile for an existing user (admin only)",
        "tags": [
          "cooks"
        ]
      }
    },
    "/api/cooks/me/dashboard": {
      "get": {
        "operationId": "CooksController_dashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cook dashboard: pending/preparing/shipped counts, monthly sales, product stock",
        "tags": [
          "cooks"
        ]
      }
    },
    "/api/cooks/{id}": {
      "get": {
        "operationId": "CooksController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get cook profile (exact coordinates only for owner/admin; public sees city label)",
        "tags": [
          "cooks"
        ]
      },
      "patch": {
        "operationId": "CooksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update cook profile (owner or admin; isActive is admin-only)",
        "tags": [
          "cooks"
        ]
      }
    },
...
    "/api/categories": {
      "get": {
        "description": "Cursor pagination (`limit`, `cursor`). Optional `search` matches name or description.",
        "operationId": "CategoriesController_findAll",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1–100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from the previous page `pagination.nextCursor`",
            "schema": {
              "example": "eyJjcmVhdGVkQXQiOiIyMDI2LTA4LTE2VDEyOjAwOjAwLjAwMFoiLCJpZCI6IjY4YWYifQ",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive partial text match (fields depend on the resource).",
            "schema": {
              "example": "maría",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List active categories",
        "tags": [
          "categories"
        ]
      },
      "post": {
        "operationId": "CategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create category (admin)",
        "tags": [
          "categories"
        ]
      }
    },
    "/api/categories/all": {
      "get": {
        "description": "Cursor pagination (`limit`, `cursor`). Optional `search` matches name or description.",
        "operationId": "CategoriesController_findAllAdmin",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1–100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from the previous page `pagination.nextCursor`",
            "schema": {
              "example": "eyJjcmVhdGVkQXQiOiIyMDI2LTA4LTE2VDEyOjAwOjAwLjAwMFoiLCJpZCI6IjY4YWYifQ",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive partial text match (fields depend on the resource).",
            "schema": {
              "example": "maría",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all categories including inactive (admin)",
        "tags": [
          "categories"
        ]
      }
    },
    "/api/categories/{id}": {
      "get": {
        "operationId": "CategoriesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get active category by id",
        "tags": [
          "categories"
        ]
      },
      "patch": {
        "operationId": "CategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update / activate-deactivate category (admin)",
        "tags": [
          "categories"
        ]
      }
    },
...
    "/api/admin/cooks/{id}": {
      "patch": {
        "operationId": "AdminController_setCookActive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetActiveDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Activate / deactivate a cook",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Market Petronio API",
    "description": "Marketplace API for cooks and customers (Pacific region).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaymentMethodDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "nequi",
              "daviplata",
              "bank_transfer"
            ]
          },
          "details": {
            "type": "string",
            "example": "Nequi 3001234567"
          },
          "isEnabled": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "type",
          "details"
        ]
      },
      "CreateCookDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "Existing user id that will own this cook profile",
            "example": "507f1f77bcf86cd799439011"
          },
          "displayName": {
            "type": "string",
            "example": "María Rodríguez"
          },
          "bio": {
            "type": "string",
            "example": "Traditional cook from the Pacific region."
          },
          "specialties": {
            "example": [
              "Seafood",
              "Coconut candies",
              "Tamales"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "publicLocation": {
            "type": "string",
            "example": "Buenaventura, Valle del Cauca"
          },
          "longitude": {
            "type": "number",
            "description": "Longitude (GeoJSON order)",
            "example": -77.0319
          },
          "latitude": {
            "type": "number",
            "description": "Latitude (GeoJSON order)",
            "example": 3.8833
          },
          "paymentMethods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentMethodDto"
            }
          },
          "contactWhatsApp": {
            "type": "string",
            "example": "3001234567"
          }
        },
        "required": [
          "userId",
          "displayName",
          "publicLocation",
          "longitude",
          "latitude"
        ]
      },
      "UpdateCookDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "specialties": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "publicLocation": {
            "type": "string"
          },
          "longitude": {
            "type": "number"
          },
          "latitude": {
            "type": "number"
          },
          "paymentMethods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentMethodDto"
            }
          },
          "contactWhatsApp": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean",
            "description": "Admin only"
          }
        }
      },
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Prepared meals"
          },
          "description": {
            "type": "string",
            "example": "Ready-to-eat traditional dishes"
          },
          "image": {
            "type": "string",
            "example": "https://res.cloudinary.com/demo/image/upload/category.jpg"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean",
            "description": "Admin activate/deactivate"
          }
        }
      },
...
      "SetActiveDto": {
        "type": "object",
        "properties": {
          "isActive": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "isActive"
        ]
      }
    }
  }
}